requirements.txt
alembic==1.4.1
aniso8601==8.0.0
attrs==19.3.0
autopep8==1.5
bcrypt==3.1.7
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
coverage==4.4.2
enum-compat==0.0.2
eventlet==0.21.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.3
Flask-Migrate==2.5.2
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SocketIO==2.9.3
Flask-SQLAlchemy==2.4.1
Flask-Testing==0.8.0
gem==0.1.12
greenlet==0.4.12
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.5.0
itsdangerous==1.1.0
Jinja2==2.11.1
jsonschema==3.2.0
Mako==1.1.2
MarkupSafe==1.1.1
psycopg2==2.8.4
pycodestyle==2.5.0
pycparser==2.20
PyJWT==1.7.1
pyrsistent==0.15.7
python-dateutil==2.8.1
python-dotenv==0.12.0
python-editor==1.0.4
python-engineio==2.0.2
python-socketio==1.8.4
pytz==2019.3
requests==2.23.0
rope==0.16.0
six==1.14.0
soupsieve==2.0
SQLAlchemy==1.3.15
tqdm==4.44.1
urllib3==1.25.8
Werkzeug==0.16.1
zipp==3.1.0
